requirements.txt
django==4.2
djangorestframework==3.14
celery==5.2.7
requests==2.28.1
redis==4.5.5